Implement importing of normals and uv coords
Now the patch_rlg() function overwrites the normals and uv coordinates of the rlg file.
Also fixed a bug where some uv coordinates would be read incorrectly. This was caused by the fact that I was reading uv coordinates from the rlg file as unsigned, which is how I thought uv coords were encoded and I'm pretty sure that's how switch toolbox does it, but it turns out they're actually signed. There are occurrences of negative uv coordinates (very rare, but exist), which previous versions of this tool would read as a really big positive number and screw up uv mapping, even if it was unnoticeable.